bhai script kaam kar gaya ek baar toh
